+++ src/audiovideocontroller/DefaultAudioVideoController.ts
@@ -18,6 +18,7 @@ import MeetingSessionStatus from '../meetingsession/MeetingSessionStatus';
 import MeetingSessionStatusCode from '../meetingsession/MeetingSessionStatusCode';
 import MeetingSessionVideoAvailability from '../meetingsession/MeetingSessionVideoAvailability';
 import DefaultPingPong from '../pingpong/DefaultPingPong';
+import PingPong from '../pingpong/PingPong';
 import DefaultRealtimeController from '../realtimecontroller/DefaultRealtimeController';
 import RealtimeController from '../realtimecontroller/RealtimeController';
 import ReconnectController from '../reconnectcontroller/ReconnectController';
@@ -63,6 +64,7 @@ import NScaleVideoUplinkBandwidthPolicy from '../videouplinkbandwidthpolicy/NSca
 import DefaultVolumeIndicatorAdapter from '../volumeindicatoradapter/DefaultVolumeIndicatorAdapter';
 import WebSocketAdapter from '../websocketadapter/WebSocketAdapter';
 import AudioVideoControllerState from './AudioVideoControllerState';
+import IdleMonitor from "../idlemonitor/IdleMonitor";
 
 export default class DefaultAudioVideoController implements AudioVideoController {
   private _logger: Logger;
@@ -74,15 +76,17 @@ export default class DefaultAudioVideoController implements AudioVideoController
   private _mediaStreamBroker: MediaStreamBroker;
   private _reconnectController: ReconnectController;
   private _audioMixController: AudioMixController;
-
+  private pingPong: PingPong;
   private connectionHealthData = new ConnectionHealthData();
   private observerQueue: Set<AudioVideoObserver> = new Set<AudioVideoObserver>();
   private meetingSessionContext = new AudioVideoControllerState();
   private sessionStateController: SessionStateController;
+  private idleMonitor: IdleMonitor;
 
   private static MIN_VOLUME_DECIBELS = -42;
   private static MAX_VOLUME_DECIBELS = -14;
   private static PING_PONG_INTERVAL_MS = 10000;
+  private static MAX_IDLE_TIME_MS = 5*60*1000;
 
   constructor(
     configuration: MeetingSessionConfiguration,
@@ -111,6 +115,7 @@ export default class DefaultAudioVideoController implements AudioVideoController
     );
     this._audioMixController = new DefaultAudioMixController();
     this.meetingSessionContext.logger = this._logger;
+    this.idleMonitor = new IdleMonitor(DefaultAudioVideoController.MAX_IDLE_TIME_MS);
   }
 
   get configuration(): MeetingSessionConfiguration {
@@ -213,16 +218,18 @@ export default class DefaultAudioVideoController implements AudioVideoController
     this.meetingSessionContext.videosToReceive = new DefaultVideoStreamIdSet();
     this.meetingSessionContext.videosPaused = new DefaultVideoStreamIdSet();
     this.meetingSessionContext.statsCollector = new DefaultStatsCollector(this, this.logger);
+    this.pingPong = new DefaultPingPong(
+      this.meetingSessionContext.signalingClient,
+      DefaultAudioVideoController.PING_PONG_INTERVAL_MS,
+      this.logger
+    );
+    this.idleMonitor.addPingPongObserver(this.pingPong);
     this.meetingSessionContext.connectionMonitor = new SignalingAndMetricsConnectionMonitor(
       this,
       this._realtimeController,
       this._videoTileController,
       this.connectionHealthData,
-      new DefaultPingPong(
-        this.meetingSessionContext.signalingClient,
-        DefaultAudioVideoController.PING_PONG_INTERVAL_MS,
-        this.logger
-      ),
+      this.pingPong,
       this.meetingSessionContext.statsCollector
     );
     this.meetingSessionContext.reconnectController = this._reconnectController;
@@ -238,9 +245,9 @@ export default class DefaultAudioVideoController implements AudioVideoController
 
     if (this._reconnectController.hasStartedConnectionAttempt()) {
       // This does not reset the reconnect deadline, but declare it's not the first connection.
-      this._reconnectController.startedConnectionAttempt(false);
+      this._reconnectController.startedConnectionAttempt(false, this.idleMonitor);
     } else {
-      this._reconnectController.startedConnectionAttempt(true);
+      this._reconnectController.startedConnectionAttempt(true, this.idleMonitor);
     }
 
     try {
@@ -492,7 +499,7 @@ export default class DefaultAudioVideoController implements AudioVideoController
 
   private async actionReconnect(): Promise<void> {
     if (!this._reconnectController.hasStartedConnectionAttempt()) {
-      this._reconnectController.startedConnectionAttempt(false);
+      this._reconnectController.startedConnectionAttempt(false, this.idleMonitor);
       this.forEachObserver(observer => {
         Maybe.of(observer.audioVideoDidStartConnecting).map(f => f.bind(observer)(true));
       });